ARTEMIS-4173 Improving reliability on RandomReattachTest
- making the test to run faster and removing the soak and stress versions as they are not really needed
- allowing expected exceptions (cannot find ACK, and making it retry the operation)
- removing duplicates of the test that are not really needed.
